Refactored pre.ChallengeResponse into fragments.CorrectnessProof
