+++ umbral/fragments.py
@@ -75,12 +75,72 @@ class KFrag(object):
         return self.to_bytes()
 
 
+class CorrectnessProof(object):
+    def __init__(self, e2, v2, u1, u2, z1, z2, z3):
+        self.point_eph_e2 = e2
+        self.point_eph_v2 = v2
+        self.point_kfrag_commitment = u1
+        self.point_kfrag_pok = u2
+        self.bn_kfrag_sig1 = z1
+        self.bn_kfrag_sig2 = z2
+        self.bn_sig = z3
+
+    @classmethod
+    def from_bytes(cls, data: bytes, curve: ec.EllipticCurve=None):
+        """
+        Instantiate CorrectnessProof from serialized data.
+        """
+        curve = curve if curve is not None else default_curve()
+        key_size = get_curve_keysize_bytes(curve)
+        data = BytesIO(data)
+
+        # BigNums are the keysize in bytes, Points are compressed and the
+        # keysize + 1 bytes long.
+        e2 = Point.from_bytes(data.read(key_size + 1), curve)
+        v2 = Point.from_bytes(data.read(key_size + 1), curve)
+        kfrag_commitment = Point.from_bytes(data.read(key_size + 1), curve)
+        kfrag_pok = Point.from_bytes(data.read(key_size + 1), curve)
+        kfrag_sig1 = BigNum.from_bytes(data.read(key_size), curve)
+        kfrag_sig2 = BigNum.from_bytes(data.read(key_size), curve)
+        sig = BigNum.from_bytes(data.read(key_size), curve)
+
+        return cls(e2, v2, kfrag_commitment, kfrag_pok, kfrag_sig1, kfrag_sig2, sig)
+
+    def to_bytes(self) -> bytes:
+        """
+        Serialize the CorrectnessProof to a bytestring.
+        """
+        e2 = self.point_eph_e2.to_bytes()
+        v2 = self.point_eph_v2.to_bytes()
+        kfrag_commitment = self.point_kfrag_commitment.to_bytes()
+        kfrag_pok = self.point_kfrag_pok.to_bytes()
+        kfrag_sig1 = self.bn_kfrag_sig1.to_bytes()
+        kfrag_sig2 = self.bn_kfrag_sig2.to_bytes()
+        sig = self.bn_sig.to_bytes()
+
+        result = e2            \
+            + v2               \
+            + kfrag_commitment \
+            + kfrag_pok        \
+            + kfrag_sig1       \
+            + kfrag_sig2       \
+            + sig
+
+        return result
+
+    def __bytes__(self):
+        return self.to_bytes()
+
+
 class CapsuleFrag(object):
     def __init__(self, e1, v1, id_, x):
         self.point_eph_e1 = e1
         self.point_eph_v1 = v1
         self.bn_kfrag_id = id_
         self.point_eph_ni = x
+
+    def attach_correctness_proof(self, proof: CorrectnessProof=None):
+        self.proof = proof
 
     @classmethod
     def from_bytes(cls, data: bytes, curve: ec.EllipticCurve = None):
@@ -98,6 +158,9 @@ class CapsuleFrag(object):
         kfrag_id = BigNum.from_bytes(data.read(key_size), curve)
         eph_ni = Point.from_bytes(data.read(key_size + 1), curve)
 
+        if data.readable():
+            proof = CorrectnessProof.from_bytes(data.read(), curve)
+
         return cls(e1, v1, kfrag_id, eph_ni)
 
     def to_bytes(self):
@@ -109,7 +172,12 @@ class CapsuleFrag(object):
         kfrag_id = self.bn_kfrag_id.to_bytes()
         eph_ni = self.point_eph_ni.to_bytes()
 
-        return e1 + v1 + kfrag_id + eph_ni
+        if self.proof:
+            proof = self.proof.to_bytes()
+
+        return e1 + v1 + kfrag_id + eph_ni + proof
 
     def __bytes__(self):
         return self.to_bytes()
+
+
